fix: Make declarations work without ts proxy lib #718
